pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Leonardo Gatti <leonardogattidev@gmail.com>"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "chatbot.main:main"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.52.2"
python-telegram-bot = "^21.6"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
